Fixed bugs and loading of galaxy files
The cut files now load... There was a bug with loading the data to memory. This has now been fixed by circumventing the error and retrying the loading of the data. Seemingly python progress whilst loading the dataset, resulting in the error.

The full processing of the hydrosim data now works as expected. It is untested with the cut files but works with the full snapshot files. Quality of Life changes for robustness have been implemented. Both cut and full simulation files should work, with centering on the galaxy applied solely to the full simulation file (uncentred by default).